tr1/output: move shader to libtrx
