+++ src/libtrx/include/libtrx/game/output/state.h
@@ -1,6 +1,12 @@
 #pragma once
 
+#include <GL/glew.h>
 #include <stdint.h>
+
+void Output_GetPerspProjectionMatrix(GLfloat output[][4]);
+void Output_GetOrthoProjectionMatrix(GLfloat output[][4]);
+
+extern int32_t Output_GetTime(void);
 
 extern void Output_SetupBelowWater(bool is_underwater);
 extern void Output_SetupAboveWater(bool is_underwater);
@@ -10,6 +16,11 @@ extern int32_t Output_GetFogEnd(void);
 
 void Output_SetFogStart(int32_t dist);
 extern void Output_SetFogEnd(int32_t dist);
+
+extern int32_t Output_GetNearZ(void);
+extern int32_t Output_GetFarZ(void);
+int32_t Output_GetNearZ_UI(void);
+int32_t Output_GetFarZ_UI(void);
 
 extern int32_t Output_GetLightAdder(void);
 extern void Output_SetLightAdder(int32_t adder);
