+++ src/tr1/game/output.h
@@ -3,11 +3,11 @@
 #include "game/output/draw.h"
 #include "game/output/func.h"
 #include "game/output/scene_compositor.h"
-#include "game/output/shader.h"
 #include "game/output/state.h"
 #include "global/types.h"
 
 #include <libtrx/game/output.h>
+#include <libtrx/game/output/shader.h>
 
 #include <stddef.h>
 #include <stdint.h>
