+++ src/tr1/game/output/types.h
@@ -3,12 +3,6 @@
 #include <libtrx/game/math/types.h>
 #include <libtrx/game/output/types.h>
 
-typedef enum {
-    LIGHTING_MODE_OFF,
-    LIGHTING_MODE_ONLY_SHADES,
-    LIGHTING_MODE_FULL,
-} LIGHTING_MODE;
-
 typedef struct {
     XYZ_32 from;
     XYZ_32 to;
